Rename to take_pill
