+++ lib/domain/record/util/take_pill.dart
